refactor: 백엔드 파일명 수정
- 컨벤션에 맞게 수정
