+++ server/src/api/routes/accountbook.js
@@ -1,6 +1,6 @@
 const Router = require('koa-router');
 const accountbookController = require('@controllers/accountbook');
-const verificationOfAuth = require('@middlewares/jwt-authentication');
+const verificationOfAuth = require('@middlewares/jwtAuthentication');
 
 const router = new Router();
